# perf: Optimize resolving X-Forwarded-For addresses