--- a/lib/request.js
+++ b/lib/request.js
@@ -50,7 +50,7 @@ function getTrustProxyFn (tp) {
   }
   if (tp === true) {
     // Support plain true/false
-    return undefined
+    return null
   }
   if (typeof tp === 'number') {
     // Support trusting hop count