--- a/lib/request.js
+++ b/lib/request.js
@@ -49,7 +49,7 @@ function getTrustProxyFn (tp) {
     return tp
   }
   if (tp === true) {
-    // Support plain true/false
+    // Support trusting everything
     return null
   }
   if (typeof tp === 'number') {