Hello, I have a suggestion. When looking up how trustProxy option in Fastify works in detail, I found out that the proxy-addr dependency used by Fastify, specifically the `.all()` function, checks if the trust argument is provided, and if it's not, then it just returns all adresses, skipping the trust checking logic alltogether, which can speed up things greatly, especially if there's a lot of X-Forwarded-For addresses. So, I made this PR, replacing the `return true` function with `undefined`, which has the same effect in the `.ips` property getter of `Request`, but because the proxy-addr default function doesn't support skipping trust argument, I replaced it with proxy-addr `.all()` function in the `.ip` property getter and did the same thing there as the proxy-addr default function does, so everything behaves exactly the same as before. If something needs further explanation please let me know.

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [ ] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)