@dmkng 2024-05-10
> can you adjust the pr title according to the rules?

I added `perf:` at the beggining, is there something else to adjust?

> how much perf does this give us?

I made a simple test that sends 11 addresses in X-Forwarded-For to the Fastify server with `trustProxy: true` and a single endpoint printing `.ip` and `.ips` properties of `Request`, I can add it to the PR or to the separate branch of my fork.

Without the optimization:

```
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1]
[1]
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 23 ms │ 26 ms │ 34 ms │ 83 ms │ 27.37 ms │ 14.94 ms │ 402 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬────────┬──────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%  │ Avg      │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼────────┼──────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 26 415  │ 26 415  │ 37 055  │ 37 503 │ 35 877,6 │ 2460,9  │ 26 404  │
[1] ├───────────┼─────────┼─────────┼─────────┼────────┼──────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 12.7 MB │ 12.7 MB │ 17.8 MB │ 18 MB  │ 17.2 MB  │ 1.18 MB │ 12.7 MB │
[1] └───────────┴─────────┴─────────┴─────────┴────────┴──────────┴─────────┴─────────┘
[1]
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1]
[1] 1077k requests in 30.05s, 516 MB read
```

With the optimization:

```
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1]
[1]
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬──────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev    │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼──────────┼────────┤
[1] │ Latency │ 22 ms │ 25 ms │ 29 ms │ 81 ms │ 26.49 ms │ 14.54 ms │ 261 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴──────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬────────┬───────────┬────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%  │ Avg       │ Stdev  │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼────────┼───────────┼────────┼─────────┤
[1] │ Req/Sec   │ 29 055  │ 29 055  │ 37 407  │ 37 631 │ 37 050,94 │ 1512,6 │ 29 044  │
[1] ├───────────┼─────────┼─────────┼─────────┼────────┼───────────┼────────┼─────────┤
[1] │ Bytes/Sec │ 13.9 MB │ 13.9 MB │ 17.9 MB │ 18 MB  │ 17.7 MB   │ 723 kB │ 13.9 MB │
[1] └───────────┴─────────┴─────────┴─────────┴────────┴───────────┴────────┴─────────┘
[1]
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1]
[1] 1112k requests in 30.05s, 531 MB read
```